Refactor - extract functionality into structs
