+++ internal/github/util.go
@@ -1,0 +1,9 @@
+package github
+
+import "log"
+
+func panicIfError(err error) {
+	if err != nil {
+		log.Fatalf("Error that we cannot handle, %v", err)
+	}
+}
